gisc-backend/pyproject.toml
[tool.poetry]
name = "app"
version = "1.0.0"
description = "TYRANTHOS TIER-0 Cyber Intelligence Operations System"
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.128.0"}
psycopg = {extras = ["binary"], version = "^3.3.2"}
sqlalchemy = "^2.0.45"
aiosqlite = "^0.22.1"
python-multipart = "^0.0.21"
pydantic-settings = "^2.12.0"
psutil = "^7.2.1"
selenium = "^4.27.1"
webdriver-manager = "^4.0.2"
beautifulsoup4 = "^4.12.3"
lxml = "^5.3.0"
requests = "^2.32.3"
pycryptodome = "^3.21.0"
reportlab = "^4.2.5"
dnspython = "^2.7.0"
scapy = "^2.7.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
slowapi = "^0.1.9"
redis = "^5.0.1"
pysocks = "^1.7.1"
yara-python = "^4.5.0"
shodan = "^1.31.0"
vt-py = "^0.18.0"
pyotp = "^2.9.0"
cryptography = "^42.0.0"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
httpx = "^0.27.0"
websockets = "^12.0"
pymisp = "^2.4.183"
OTXv2 = "^1.5.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
